better code reorganization for the clamp patch
